refactor: move StringUtils package to step1.util package
